S3 key generation: dev workaround
There seems to be an issue with the key generation where it puts a forward-slash at the start of the key. This is a slightly hacky change which cleans the slash from the key.

The issue is probably being caused by a Java or Scala version issue with `mkstring` or `String.split`. More investigation might provide a better fix.