Sidebar: styling changes
This swaps the timestamp and the relative date around as the timestamp is the more valuable information.

It also makes the interval between snapshots always visible but less intrusive. This is a minor detail at the moment but when event driven snapshots start happening this information is more relevant and inconvenient to access with a hover over.