Adopt guardian/setup-scala to fix CI
The CI is currently broken on this repository because it expects sbt to
be preinstalled on the runners but ubuntu-latest now points to an
ubuntu-24.04 build that doesn’t include sbt (e.g. see [this
build](https://github.com/guardian/flexible-restorer/actions/runs/13160341152/job/36727220402?pr=110)).

To fix CI, this commit adopts
[guardian/setup-scala](https://github.com/guardian/setup-scala) to
install sbt on the runners and configure it how we usually do.

To use setup-scala, a .tool-versions file specifying the desired version
of Java is required: this commit adds this pointing to a Java 11 version
of corretto, which this app was already using.