Support CODE and DEV tracking pixel
This commit updates the use of the tools audit tracking pixel to use the
CODE and DEV telemetry backends when this app is being run at the
corresponding stage, and to omit the pixel when being run at an
unrecognised stage.

I’m a little uncertain whether it should use the DEV or CODE telemetry
backend when restorer is running in DEV, but I figure DEV is the safer
choice.