Use tracking pixel only in angular app
Including the tracking pixel in both the scala code and the angular app
means having it in two places at once, and doesn’t seem to provide much
benefit over just having it in the angular app.

This commit removes the tracking pixel from the scala code and updates
the angular use to track on initial page load as well as on subsequent
navigations.

This reverts commit fed80dc90befdd28c0535149352a9a4c6611e89c and commit
d6e256fb5925e9ca146092505aa8bda92c531843.